Convert internal mem_taken type to size_t.

+++ code/bmpman/bm_internal.h
@@ -63,7 +63,7 @@ struct bitmap_entry {
 	int dir_type;           //!< which directory this was loaded from (to skip other locations with same name)
 
 	// compressed bitmap stuff (.dds) - RT please take a look at this and tell me if we really need it
-	int mem_taken;          //!< How much memory does this bitmap use? - UnknownPlayer
+	size_t mem_taken;          //!< How much memory does this bitmap use? - UnknownPlayer
 	int num_mipmaps;        //!< number of mipmap levels, we need to read all of them
 
 	// Stuff to keep track of usage
